rename Prayer to PrayerRequest
