+++ packages/apollos-data-connector-rock/src/index.js
@@ -15,7 +15,7 @@ import * as Utils from './utils';
 import * as BinaryFiles from './binary-files';
 import * as Feature from './features';
 import * as Event from './events';
-import * as Prayer from './prayer';
+import * as PrayerRequest from './prayer-request';
 
 export {
   Followings,
@@ -35,5 +35,5 @@ export {
   BinaryFiles,
   Feature,
   Event,
-  Prayer,
+  PrayerRequest,
 };
